Tokenizer Checker functions
